Make worker threads retry fetching a valid task before quitting
Sometime it takes some time for the initial task to spawn child tasks.
(eg: when the initial task has to download a file first). In that case,
other worker threads should not quit because they couldn't find a task
to work on.

When running in non-blocking mode, a worker thread will retry to get a task
a bunch of time. It will exit only after a retry limit is reached.