Misc Promethus changes (#130)
* Remove default GC and process collectors

* Use Prometheus metrics prefix

Even though it is considered an anti-pattern to add a prefix with the name of the software or component to metrics (according to the official Prometheus documentation), I have decided to add a prefix. I’ve found that this makes it much easier to find relevant metrics. The main disadvantage of per-component prefixes queries become slightly more complex if you want to query the same metric (e.g. HTTP request duration) across multiple components. This isn’t super important in our case though, so I think the trade-off is acceptable.

* Use keyword args to set Prometheus metric labels

As suggested by @stchris, thanks!