Standardize development dependencies / refactor GHA workflow (#153)
* Standardize development dependencies

There seems to be a preference towards using `requirements-dev.txt` vs. a `dev` extra for development dependencies, but right now neither `pip install -r requirements-dev.txt` nor `pip install -e ".[dev]" is enough to install all development dependencies.

I’ve added any dependencies listed in the `dev` extra to `requirements-dev.txt` and removed the `dev` extra so that we only have one list of dev dependencies to maintain.

This also pins the `moto` package (used to mock S3 in tests) to version 4.x as version 5 release a few days ago contains breaking changes.

* Remove unnecessary setup in GitHub Actions workflow

This seems to be left over from when we still used to run tests directly in GHA runners.

* Install servicelayer before running tests

Tests in `test_extensions.py` rely on `servicelayer` being installed as they ultimately read information from the egg info that is created during installation.

This is basically equivalent to the previous setup which executed `pip install -e .` outside of the container. As the entire directory is mounted into the container, the egg info subsequently was also available inside of the container. While that worked I found the fact that installing something outside of the container could make the tests fail or pass quite confusing. This should be a little more explicit.

* Run everything inside of the docker container

Running some steps outside and some inside of the container leads to weird behavior, e.g. when file permissions/owners do not match across steps.

* Slim down Docker image

This uses a pre-built Python image based on Debian as this should be much faster than building a custom image on top of a full-feature Ubuntu.

* Add separate Makefile target to run tests without starting a Docker container