Fix dataset timestamps (#190)
* Reproduce incorrect dataset timestamps

https://github.com/alephdata/aleph/issues/3787

* Remove "end_time" timestamp

We currently do not retain information about inactive datasets (i.e. datasets that do not have any more pending or running tasks). For this reason, the "end_time" timestamp is currently of no use, as it would never be displayed to users anyway.

While there are some plans around providing more detail about the results of processed tasks as well as completed jobs, it is unclear where this data will be stored and what the implementation will look like. As it is easy enough to add this information back (< 10 LOC), I’ve removed it for now.

* Only set `start_time` timestamp if it hasn’t been set yet

* Delete dataset timestamps when all tasks have been processed

* Add tests covering dataset status when cancelling tasks

* Extract flushing status data into separate method

* Ensure time-machine is installed in GitHub Actions

* Delete dead code

* Remove redundant code

`cleanup_dataset_status` iterates over all active datasets and removes status information if it is done (i.e. there are no more pending or running tasks). It is called by the Aleph worker periodically.

We already do this for individual datasets whenever a task from the dataset is done or the dataset is cancelled as a whole. That means that `cleanup_dataset_status` is redundant.

Also see: https://github.com/alephdata/servicelayer/pull/190#issuecomment-2233845128

* Use `is_done` helper method

* Fix linter errors

---------

Co-authored-by: catileptic <alex.stefanescu@protonmail.com>